+++ src/main/java/com/chaitanya/company/model/CompanyDTO.java
@@ -6,14 +6,14 @@ public class CompanyDTO extends BaseDTO {
 
 	private static final long serialVersionUID = 1L;
 	
-	private Integer companyId;
+	private Long companyId;
 	private String companyCode;
 	private String companyName;
 	
-	public Integer getCompanyId() {
+	public Long getCompanyId() {
 		return companyId;
 	}
-	public void setCompanyId(Integer companyId) {
+	public void setCompanyId(Long companyId) {
 		this.companyId = companyId;
 	}
 	public String getCompanyCode() {
